com.eastapps.meme_gen_android.test.MemeServerClientTest.testStoreMemeUser now working

+++ meme_gen_android/src/com/eastapps/meme_gen_android/web/IMemeServerClient.java
@@ -18,7 +18,7 @@ public interface IMemeServerClient {
 
 	List<MemeBackground> getPopularMemeBackgrounds(final String popularTypeName);
 
-	int storeNewUser(MemeUser shallowUser);
+	long storeNewUser(MemeUser shallowUser);
 
 	List<MemeBackground> getAllMemeBackgrounds();
